requirements.txt
Django==1.6.5
Pillow==2.5.3
South==0.8.4
django-classy-tags==0.5
django-cms==3.0
django-mptt==0.6.0
django-reversion==1.8.2
django-sekizai==0.7
djangocms-admin-style==0.2.2
djangocms-column==1.4
djangocms-grid==1.0.5
djangocms-link==1.4.0
djangocms-oembed==0.4.1
djangocms-snippet==1.1
djangocms-style==1.4
djangocms-table==1.1.0
djangocms-text-ckeditor==2.3.0
html5lib==1.0b1
micawber==0.3.0
pyquery==1.2.9
six==1.3.0
yolk==0.4.3
djangocms-file
